breaking into smaller pieces
