+++ network/dispatcher.py
@@ -1,0 +1,14 @@
+from minimalog.minimal_log import MinimalLog
+ml = MinimalLog(__name__)
+
+
+class Operator:
+    def __init__(self):
+        pass
+
+    @staticmethod
+    def transfer_files_to_remote():
+        try:
+            pass
+        except Exception as e_err:
+            ml.log_event(e_err, level=ml.ERROR)
