+++ qbit_interface/api_helper.py
@@ -1,85 +1,10 @@
-from configparser import ConfigParser  # only used to type a return value
-from configuration_reader import get_user_configuration
 from datetime import datetime
 from minimalog.minimal_log import MinimalLog
-from re import findall
-from time import sleep
-from user_configuration.WEB_API_CREDENTIALS import HOST, USER, PASS
-# TODO delete 'import inspect' if remains unused, could be useful?
-import inspect
-import qbittorrentapi
+ml = MinimalLog(__name__)
 
 
-class QbitTasker:
-    def __init__(self, manage_log_files=False):
-        ml.log_event('initialize \'{}\''.format(self.__class__), event_completed=False, announce=True)
-        self.config = get_user_configuration()
-        assert self.config is not None, ml.log_event('!! no user configuration !!', announce=True, level=ml.ERROR)
-        self.main_loop_count = 0
-        self.qbit_client_connected = True if self._client_is_connected() else False
-        self._connection_time_start = datetime.now()
-        self._reset_search_ids()
-        self.active_search_ids, self.active_header = dict(), ''
-        if manage_log_files:
-            self._manage_log_files()  # TODO this does nothing
-        ml.log_event('initialize \'{}\''.format(self.__class__), event_completed=True, announce=True)
-        self.pause_on_event(self.config.hardcoded.keys.user_config_parser_keyring.WAIT_FOR_USER)
-
-    def increment_loop_count(self):
-        self.main_loop_count += 1
-        ml.log_event('current connection to qbittorrent api was started at {}'.format(self._connection_time_start))
-        ml.log_event('main loop has ended, {} total loops..'.format(self.main_loop_count))
-        ml.log_event('when in doubt, compare parsed file keys with config reader string values', level=ml.WARNING)
-        ml.log_event('when in doubt, compare parsed file keys with config reader string values', level=ml.WARNING)
-
-    def initiate_and_monitor_searches(self):
-        try:
-            # FYI cannot use self._get_parser functions here because self.active_headers has not been set
-            search_detail_parser_section_headers = self.config.parser.parsers.search_detail_parser.sections()
-            user_config_parser = self._get_user_config_parser()
-            search_detail_parser_keys = self.config.hardcoded.keys.search_parser_keyring
-            user_config_parser_keys = self._get_keyring_for_user_config_parser()
-            # perform sorting on search queue
-            sort_key = search_detail_parser_keys.TIME_LAST_SEARCHED
-            self._set_search_order_ranking_by_(sort_key)
-            self.pause_on_event(user_config_parser_keys.WAIT_FOR_USER)
-            # TODO reminder, main program loop entry here TODO $
-            for search_detail_parser_section_header in search_detail_parser_section_headers:
-                self.active_header = search_detail_parser_section_header
-                ml.log_event(f'monitoring search header \'{self.active_header}\'')
-                self._manage_state_updates(self._get_search_state())
-            self._config_to_disk()
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def pause_on_event(self, pause_type):
-        try:
-            timestamp = _get_timestamp()
-            user_config_parser_keys = self._get_keyring_for_user_config_parser()
-            user_config_parser = self.config.parser.parsers.user_config_parser
-            user_config_parser_at_default = user_config_parser[user_config_parser_keys.DEFAULT]
-            if pause_type == user_config_parser_keys.WAIT_FOR_MAIN_LOOP:
-                delay = int(user_config_parser_at_default[user_config_parser_keys.WAIT_FOR_MAIN_LOOP])
-                ml.log_event('{} waiting {} seconds for main loop repeat..'.format(timestamp, delay))
-                sleep(delay)
-            elif pause_type == user_config_parser_keys.WAIT_FOR_SEARCH_STATUS_CHECK:
-                delay = int(user_config_parser_at_default[user_config_parser_keys.WAIT_FOR_SEARCH_STATUS_CHECK])
-                ml.log_event('{} waiting {} seconds for search state check..'.format(timestamp, delay))
-                sleep(delay)
-            elif pause_type == user_config_parser_keys.WAIT_FOR_SEARCH_RESULT_ADD:
-                delay = int(user_config_parser_at_default[user_config_parser_keys.WAIT_FOR_SEARCH_RESULT_ADD])
-                ml.log_event('{} waiting {} seconds for add attempt..'.format(timestamp, delay))
-                sleep(delay)
-            else:
-                delay = int(user_config_parser_at_default[user_config_parser_keys.WAIT_FOR_USER])
-                # TODO miscellaneous is hardcoded, not that it matters, just annoying and causes upkeep
-                ml.log_event('{} waiting {} seconds to let user follow log..'.format(timestamp, delay))
-                sleep(delay)
-            # ml.log_event('\n')  # puts one empty line after pauses, for visual affect in the log
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def transfer_files_to_remote(self):
+class QbitBotHelper:
+    def __init__(self):
         pass
 
     def _active_header_search_id_is_valid(self) -> bool:
@@ -87,6 +12,15 @@ class QbitTasker:
             if self.active_header in self.active_search_ids:
                 return True
             return False
+        except Exception as e_err:
+            ml.log_event(e_err, level=ml.ERROR)
+
+    def _check_if_search_is_concluded(self):
+        try:
+            search_parser_keys = self._get_keyring_for_search_detail_parser()
+            if self._search_has_yielded_the_required_results():
+                ml.log_event('search \'{}\' has concluded, disabling'.format(self.active_header), announce=True)
+                self._update_search_states(search_parser_keys.CONCLUDED)
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
@@ -105,38 +39,10 @@ class QbitTasker:
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
-    def _check_if_search_is_concluded(self):
-        try:
-            search_parser_keys = self._get_keyring_for_search_detail_parser()
-            if self._search_has_yielded_the_required_results():
-                ml.log_event('search \'{}\' has concluded, disabling'.format(self.active_header), announce=True)
-                self._update_search_states(search_parser_keys.CONCLUDED)
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
     def _check_if_section_header_metadata_exists_as_local_result(self) -> bool:
         delete_me = self.config  # suppress @staticmethod warning
         ml.log_event('TODO : ..just return hardcoded False..', level=ml.WARNING)
         return False  # TODO
-
-    def _client_is_connected(self) -> bool:
-        """
-        connect to the client, fetch check app version and web api version
-        :return: bool, true if able to populate all data successfully
-        """
-        ml.log_event('connect to client', event_completed=False)
-        try:
-            self.qbit_client = qbittorrentapi.Client(host=HOST, username=USER, password=PASS)
-            app_version = self.qbit_client.app_version
-            web_api_version = self.qbit_client.app_web_api_version
-            if app_version is not None and web_api_version is not None:
-                # TODO could unpack more details here just for readability
-                ml.log_event('connect to client with.. \n\n\tclient app version {} \n\tweb api version {}\n\n'.format(
-                    app_version, web_api_version), event_completed=True)
-                return True
-            return False
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
 
     def _config_to_disk(self):
         ml.log_event('writing parser configurations to disk')
@@ -225,8 +131,7 @@ class QbitTasker:
         expected_search_result_count = \
             int(search_detail_parser_at_active_header[search_detail_parser_keys.EXPECTED_SEARCH_RESULT_COUNT])
         # TODO why is expected_search_result_count == 0?
-        ml.log_event('get most popular results up to count {}'.format(
-            expected_search_result_count), event_completed=False)
+        ml.log_event(f'get most popular results up to count {expected_search_result_count}', event_completed=False)
         # TODO BUG happens here : '<' not supported between instances of 'int' and 'str'
         found_result_count = len(regex_filtered_results)
         if not _enough_results_in_(regex_filtered_results, expected_search_result_count):
@@ -325,24 +230,6 @@ class QbitTasker:
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
-    def _get_search_state(self) -> tuple:
-        """
-        :return: search states
-        """
-        try:
-            search_parser_keys = self._get_keyring_for_search_detail_parser()
-            search_detail_parser_at_active_header = self._get_search_detail_parser_at_active_header()
-            search_detail_parser_at_active_header[search_parser_keys.TIME_LAST_READ] = str(datetime.now())
-            _search_queued = search_detail_parser_at_active_header.getboolean(search_parser_keys.QUEUED)
-            _search_running = search_detail_parser_at_active_header.getboolean(search_parser_keys.RUNNING)
-            _search_stopped = search_detail_parser_at_active_header.getboolean(search_parser_keys.STOPPED)
-            _search_concluded = search_detail_parser_at_active_header.getboolean(search_parser_keys.CONCLUDED)
-            ml.log_event('search state for \'{}\': \n\tqueued: {}\n\trunning: {}\n\tstopped: {}\n\tconcluded: {}'.format(
-                self.active_header, _search_queued, _search_running, _search_stopped, _search_concluded), announce=True)
-            return _search_queued, _search_running, _search_stopped, _search_concluded
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
     def _get_search_term_from_search_detail_parser_at_active_header(self) -> str:
         ml.log_event('get search term for search detail parser at header \'{}\''.format(self.active_header))
         search_parser_keys = self._get_keyring_for_search_detail_parser()
@@ -418,63 +305,22 @@ class QbitTasker:
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
-    def _manage_state_updates(self, section_states):
-        try:
-            ml.log_event('manage state updates..')
-            _search_queued, _search_running, _search_stopped, _search_concluded = section_states
-            # abbreviate class objects
-            search_detail_parser_at_active_header = self._get_search_detail_parser_at_active_header()
-            search_detail_parser_keys = self.config.hardcoded.keys.search_parser_keyring
-            user_config_parser_keys = self.config.hardcoded.keys.user_config_parser_keyring
-            # search sorting
-            search_rank = \
-                int(search_detail_parser_at_active_header[search_detail_parser_keys.SEARCH_RANK])
-            # perform the appropriate action based on the search parser's values
-            # TODO should i force search rank to be 0 to only allow the top prioritized?
-            if _search_queued and not self._search_queue_full() and search_rank < 3:  # TODO un-hardcode this
-                self._start_search()  # search is in queue and queue has room, attempt to start this search
-            elif _search_running:
-                _search_status = self._qbit_get_search_status()
-                if search_detail_parser_keys.RUNNING in _search_status:
-                    pass  # search is ongoing, do nothing
-                elif search_detail_parser_keys.STOPPED in _search_status:
-                    self._update_search_states(search_detail_parser_keys.STOPPED)  # mark search as stopped (finished)
-                else:
-                    self._update_search_states(search_detail_parser_keys.QUEUED)  # search status unexpected, re-queue
-            elif _search_stopped:
-                regex_filtered_results, regex_filtered_results_count = self._get_regex_filtered_results_and_count()
-                if regex_filtered_results is not None and regex_filtered_results_count > 0:
-                    # TODO results_key.supply could be sort by any key, how to get that value here?
-                    search_priority = self._get_priority_key_for_search_result_sorting()
-                    self._save_remote_metadata_to_local_results_sorting_by_(
-                        search_priority, regex_filtered_results)  # search is finished, attempt to add results
-                else:
-                    ml.log_event(f're-queueing search for {self.active_header}..')
-                    self._update_search_states(search_detail_parser_keys.QUEUED)  # no results found, re-queue
-            elif _search_concluded:
-                pass
-            else:
-                self._update_search_states(search_detail_parser_keys.QUEUED)
-            self.pause_on_event(user_config_parser_keys.WAIT_FOR_SEARCH_STATUS_CHECK)
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
     def _metadata_parser_write_to_metadata_config_file(self, result):
         try:
             metadata_parser_keys, user_config_parser_keys = \
                 self._get_keyring_for_metadata_parser(), self._get_keyring_for_user_config_parser()
-            ml.log_event('save metadata result to file: {}'.format(result[metadata_parser_keys.NAME]))
+            ml.log_event(f'save metadata result to file: {result[metadata_parser_keys.NAME]}')
             metadata_section = self._hash(result[metadata_parser_keys.NAME])
             if not self.config.parser.parsers.metadata_parser.has_section(metadata_section):
-                ml.log_event('qbit client has added result \'{}\' for header \'{}\''.format(
-                    result[metadata_parser_keys.NAME], self.active_header), announce=True)
+                ml.log_event(f'qbit client has added result \'{result[metadata_parser_keys.NAME]}\' for header'
+                             f' \'{self.active_header}\'', announce=True)
                 self.config.parser.parsers.metadata_parser.add_section(metadata_section)
                 header = metadata_section
                 for attribute, detail in result.items():
                     # TODO there are some redundant log commands 'above' and 'below' this entry
                     # TODO i think this entry is causing the redundant log commands with _hash() calls
                     h_attr, d_attr = self._hash(attribute), self._hash(detail)
-                    ml.log_event('detail added to metadata parser with attribute key \'{}\''.format(h_attr))
+                    ml.log_event(f'detail added to metadata parser with attribute key \'{h_attr}\'')
                     self.config.parser.parsers.metadata_parser[header][h_attr] = d_attr
                     self.pause_on_event(user_config_parser_keys.WAIT_FOR_USER)
         except Exception as e_err:
@@ -484,7 +330,6 @@ class QbitTasker:
         try:
             search_detail_parser_at_active_header = self._get_search_detail_parser_at_active_header()
             search_detail_parser_keys = self._get_keyring_for_search_detail_parser()
-            user_config_parser = self._get_user_config_parser()
             user_config_parser_keys = self._get_keyring_for_user_config_parser()
             primary_search_term = search_detail_parser_at_active_header[search_detail_parser_keys.PRIMARY_SEARCH_TERM]
             regex_match = findall(filename_regex, filename)
@@ -495,98 +340,6 @@ class QbitTasker:
                 self.pause_on_event(user_config_parser_keys.WAIT_FOR_USER)
                 return True
             return False
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def _qbit_add_result(self, result):
-        try:
-            search_parser_keys, user_config_parser_keys, metadata_parser_keys = \
-                self.config.hardcoded.keys.search_parser_keyring, \
-                self.config.hardcoded.keys.user_config_parser_keyring, \
-                self.config.hardcoded.keys.metadata_parser_keyring
-            search_detail_parser_at_active_header = self._get_search_detail_parser_at_active_header()
-            count_before = self._qbit_count_all_torrents()
-            ml.log_event('local machine has {} stored results before add attempt..'.format(count_before))
-            # TODO why does this api call sometimes not add? bad result? not long enough wait?
-            # self.qbit_client.torrents_add(urls=result['fileUrl'], is_paused=True)  # TODO delete
-            self.qbit_client.torrents_add(urls=result[metadata_parser_keys.URL], is_paused=True)
-            self.pause_on_event(user_config_parser_keys.WAIT_FOR_SEARCH_RESULT_ADD)
-            results_added = self._qbit_count_all_torrents() - count_before
-            # TODO why does client fail to add so much? async opportunity? bad results? dig into api code perhaps
-            if results_added > 0:  # successful add
-                self._metadata_parser_write_to_metadata_config_file(result)
-                search_detail_parser_at_active_header[search_parser_keys.RESULTS_ADDED_COUNT] = \
-                    str(int(search_detail_parser_at_active_header[search_parser_keys.RESULTS_ADDED_COUNT]))
-                return
-            ml.log_event('client failed to add \'{}\''.format(result[metadata_parser_keys.NAME]), level=ml.WARNING)
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def _qbit_count_all_torrents(self) -> int:
-        try:
-            local_result_count = 0
-            if self.qbit_client.torrents.info().data:
-                local_result_count = len(self.qbit_client.torrents.info().data)
-            return local_result_count
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def _qbit_create_search_job(self, pattern, plugins, category) -> tuple:
-        try:
-            _job = self.qbit_client.search.start(pattern, plugins, category)
-            _status = _job.status()
-            _state = _status.data[0]['status']
-            _id = str(_status.data[0]['id'])
-            _count = _status.data[0]['total']
-            ml.log_event('qbit client created search job for \'{}\''.format(pattern))
-            return _job, _status, _state, _id, _count
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def _qbit_get_search_results(self):
-        try:
-            # FIXME this needs to handle invalid search id values.. such as values from previous runs
-            search_id = self._get_active_search_id()
-            if search_id:
-                results = self.qbit_client.search_results(search_id)
-                if results is None:
-                    ml.log_event(f'results for header \'{self.active_header}\' with '
-                                 f'search id \'{search_id}\' is None', level=ml.WARNING)
-                ml.log_event('qbit client get search results', event_completed=True)
-                return results
-            return None
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
-    def _qbit_get_search_status(self) -> str:
-        """
-        :return: the status of the search at search_id
-        """
-        try:
-            ml.log_event(f'check search status for section \'{self.active_header}\'')
-            search_parser_keys = self.config.hardcoded.keys.search_parser_keyring
-            search_detail_parser_at_active_header = self._get_search_detail_parser_at_active_header()
-            search_id, search_status = search_detail_parser_at_active_header[search_parser_keys.SEARCH_ID], None
-            # TODO did this line below add any value? delete this line after decision
-            # search_id_valid = self._active_header_search_id_is_valid()  # FIXME, i think i deleted this function?
-            # TODO PRIORITY BUG, this section of code was double logging, fixed?
-            ml.log_event(f'getting search status for header \'{self.active_header}\' with search id \'{search_id}\'')
-            if search_id in self.active_search_ids.values():
-                ongoing_search = self.qbit_client.search_status(search_id=search_id)
-                if ongoing_search is None:
-                    ml.log_event(f'are you sure that search id \'{search_id}\' is valid?')
-                    ml.log_event(f'known valid search ids are..')
-                    for search_header, search_id in self.active_search_ids:
-                        ml.log_event(f'\t header \'{search_id}\' with id \'{search_id}\'')
-                    pass  # FIXME
-                # FIXME ASAP line above is causing a NoneType return, soft-freezing the state-machine
-                search_status = ongoing_search.data[0]['status']
-            if search_status is None:  # TODO fyi new line, monitor, delete comment after
-                ml.log_event(f'search status is \'{search_status}\' for section \'{self.active_header}\'',
-                             level=ml.WARNING)
-                return search_status
-            ml.log_event(f'search status is \'{search_status}\' for section \'{self.active_header}\'')
-            return search_status
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
@@ -747,26 +500,6 @@ class QbitTasker:
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
-    def _search_queue_full(self) -> bool:
-        ml.log_event('check search queue..')
-        try:
-            if self._search_at_active_header_is_already_started():
-                return True  # TODO, add this feature
-            active_search_count = len(self.active_search_ids.keys())
-            if active_search_count < 5:
-                ml.log_event('search queue is NOT full..')
-                ml.log_event('active search headers are..')
-                for active_search_header_name in self.active_search_ids.keys():
-                    ml.log_event('search header : \'{}\''.format(active_search_header_name))
-                return False
-            ml.log_event('search queue is FULL, cannot add header \'{}\''.format(self.active_header, announce=True))
-            ml.log_event('active search headers are..')
-            for active_search_header_name in self.active_search_ids.keys():
-                ml.log_event('search header : \'{}\''.format(active_search_header_name))
-            return True
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
     def _search_set_end_reason(self, reason):
         try:
             search_parser_keys = self.config.hardcoded.keys.search_parser_keyring
@@ -852,46 +585,6 @@ class QbitTasker:
         except Exception as e_err:
             ml.log_event(e_err, level=ml.ERROR)
 
-    def _update_search_states(self, api_state_key):
-        ml.log_event('updating the search state machine..')
-        try:
-            search_parser_keys = self._get_keyring_for_search_detail_parser()
-            search_detail_parser_at_active_header = self._get_search_detail_parser_at_active_header()
-            if api_state_key == search_parser_keys.QUEUED:
-                search_detail_parser_at_active_header[search_parser_keys.QUEUED] = search_parser_keys.YES
-                search_detail_parser_at_active_header[search_parser_keys.RUNNING] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.STOPPED] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.CONCLUDED] = search_parser_keys.NO
-                self.config.parser.parsers.search_detail_parser.remove_section(search_parser_keys.SEARCH_ID)
-                ml.log_event('search for \'{}\' is queued, will be started when search queue has vacancy'.format(
-                    self.active_header))
-            elif api_state_key == search_parser_keys.RUNNING:
-                search_detail_parser_at_active_header[search_parser_keys.QUEUED] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.RUNNING] = search_parser_keys.YES
-                search_detail_parser_at_active_header[search_parser_keys.STOPPED] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.CONCLUDED] = search_parser_keys.NO
-                self._increment_search_attempt_count()
-                ml.log_event('search for \'{}\' is running.. please stand by..'.format(self.active_header))
-            elif api_state_key == search_parser_keys.STOPPED:
-                search_detail_parser_at_active_header[search_parser_keys.QUEUED] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.RUNNING] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.STOPPED] = search_parser_keys.YES
-                search_detail_parser_at_active_header[search_parser_keys.CONCLUDED] = search_parser_keys.NO
-                ml.log_event('search for \'{}\' is stopped and will be processed on next program loop'.format(
-                    self.active_header))
-            elif api_state_key == search_parser_keys.CONCLUDED:
-                search_detail_parser_at_active_header[search_parser_keys.QUEUED] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.RUNNING] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.STOPPED] = search_parser_keys.NO
-                search_detail_parser_at_active_header[search_parser_keys.CONCLUDED] = search_parser_keys.YES
-                ml.log_event('search for \'{}\' has concluded due to exceeding user preferences, '
-                             'no further action to be taken'.format(self.active_header))
-            else:
-                pass
-            search_detail_parser_at_active_header[search_parser_keys.TIME_LAST_WRITTEN] = str(datetime.now())  # TODO what???
-        except Exception as e_err:
-            ml.log_event(e_err, level=ml.ERROR)
-
     def _write_list_of_dict_to_parser(self, list_of_dicts: dict, parser: ConfigParser) -> None:
         try:
             for _dict in list_of_dicts:
@@ -920,10 +613,3 @@ def _enough_results_in_(filtered_results, expected_result_count):
 
 def _get_timestamp():
     return datetime.now()
-
-
-if __name__ == '__main__':
-    ml = MinimalLog()
-    ml.log_event('this should not be run directly, user main loop')
-else:
-    ml = MinimalLog(__name__)
